Refactor: "models" folder renamed to "schemas". Added new subschemas as interfaces for code reusing. Reordered some methods on predictions controller
