--- src/models/sumarization.ts
@@ -1,30 +1,0 @@
-export {};
-
-import {
-    Schema,
-    model
-} from 'mongoose';
-
-const sumarizationSchema = new Schema({
-    _id: Schema.Types.ObjectId,
-    cityId: Number,
-    date: Number,
-    ranges: [{
-        // TODO: Por lo visto tengo que sacar este campo date
-        date: Number,
-        start: {
-            lat: Number,
-            lng: Number
-        },
-        end: {
-            lat: Number,
-            lng: Number
-        },
-        score: Number,
-        accuracy: Number
-    }],
-});
-
-const Sumarization = model('Sumarizations', sumarizationSchema, 'sumarizations');
-
-export default Sumarization;
