+++ src/schemas/city.ts
@@ -4,15 +4,15 @@ import {
     Schema,
     model
 } from 'mongoose';
+import { 
+    iCoordinate
+} from './subschemas/coordinate';
 
 const citySchema = new Schema({
     _id: Schema.Types.ObjectId,
     id: Number,
     name: String,
-    center: {
-        lat: Number,
-        lng: Number
-    }
+    center: iCoordinate
 });
 
 const City = model('City', citySchema, 'cities', false);
