+++ src/schemas/reparation.ts
@@ -4,19 +4,16 @@ import {
     Schema,
     model
 } from 'mongoose';
+import { 
+    iCoordinate
+} from './subschemas/coordinate';
 
 const reparationSchema = new Schema({
     _id: Schema.Types.ObjectId,
     cityId: Number,
     date: Number,
-    from: {
-        lat: Number,
-        lng: Number
-    },
-    to: {
-        lat: Number,
-        lng: Number
-    },
+    from: iCoordinate,
+    to: iCoordinate,
 });
 
 const Reparation = model('Reparations', reparationSchema, 'reparations');
