+++ src/schemas/subschemas/accelerometer.ts
@@ -1,0 +1,15 @@
+export {};
+
+import { 
+    iDelta
+} from "./delta";
+
+export const iAccelerometer = {
+    id: Number,
+    eventId: Number,
+    currentTime: Number,
+    x: iDelta,
+    y: iDelta,
+    z: iDelta,
+    axis: String
+};
