+++ src/schemas/subschemas/coordinate.ts
@@ -1,0 +1,6 @@
+export {};
+
+export const iCoordinate = {
+    lat: Number,
+    lng: Number
+};
