+++ src/schemas/subschemas/delta.ts
@@ -1,0 +1,7 @@
+export {};
+
+export const iDelta = {
+    raw: Number,
+    delta: Number,
+    diff: Number
+};
