+++ src/schemas/subschemas/range.ts
@@ -1,0 +1,17 @@
+export {};
+
+import { 
+    iCoordinate
+} from "./coordinate";
+import { 
+    iStabilityEvent
+} from "./stabilityEvent";
+
+export const iRange = {
+    start: iCoordinate,
+    end: iCoordinate,
+    date: Number,
+    score: Number,
+    accuracy: Number,
+    stabilityEvents: [iStabilityEvent]
+};
