+++ src/schemas/subschemas/stabilityEvent.ts
@@ -1,0 +1,12 @@
+export {};
+
+export const iStabilityEvent = {
+    id: Number,
+    startTime: Number,
+    endTime: Number,
+    xavg: Number,
+    yavg: Number,
+    zavg: Number,
+    score: Number,
+    duration: Number
+};
